refactor: DTO 파일 프로퍼티 파싱
dto 파일 이름 컨벤션 통일 kebab-case
