+++ server/src/auth/dto/oauth-naver.dto.ts
@@ -1,4 +1,9 @@
+import { IsNotEmpty } from 'class-validator';
+
 export class OauthNaverDto {
+  @IsNotEmpty()
   authorizationCode: string;
+
+  @IsNotEmpty()
   state: string;
 }
