+++ server/src/auth/dto/user-register.dto.ts
@@ -1,16 +1,14 @@
-import { IsNotEmpty } from 'class-validator';
+import { IsEnum, IsNotEmpty } from 'class-validator';
 import { OauthInfo } from '../model/oauth-info.enum';
 
 export class RegisterUserDto {
   @IsNotEmpty()
   email: string;
 
-  // @IsNotEmpty()
-  // image: Express.Multer.File;
-
   @IsNotEmpty()
   userName: string;
 
   @IsNotEmpty()
+  @IsEnum(OauthInfo)
   oauthInfo: OauthInfo;
 }
