+++ server/src/board/dto/create-board.dto.ts
@@ -1,0 +1,30 @@
+import { Transform } from 'class-transformer';
+import { IsBoolean, IsInt, IsNotEmpty, IsOptional } from 'class-validator';
+
+export class CreateBoardDto {
+  @IsNotEmpty()
+  @IsInt()
+  @Transform(({ value }) => parseInt(value, 10))
+  userId: number;
+
+  @IsNotEmpty()
+  content: string;
+
+  @IsNotEmpty()
+  @IsBoolean()
+  @Transform(({ value }) => value === 'true')
+  isStreet: boolean;
+
+  @IsOptional()
+  location: string;
+
+  @IsOptional()
+  @IsInt()
+  @Transform(({ value }) => parseInt(value, 10))
+  latitude: number;
+
+  @IsOptional()
+  @IsInt()
+  @Transform(({ value }) => parseInt(value, 10))
+  longitude: number;
+}
