--- server/src/board/dto/createBoardDto.ts
@@ -1,18 +1,0 @@
-import { IsNotEmpty } from 'class-validator';
-
-export class CreateBoardDto {
-  @IsNotEmpty()
-  userId: number;
-
-  @IsNotEmpty()
-  content: string;
-
-  @IsNotEmpty()
-  isStreet: boolean;
-
-  location: string | null;
-
-  latitude: number;
-
-  longitude: number;
-}
