--- server/src/board/dto/deleteBoardDto.ts
@@ -1,9 +1,0 @@
-import { IsNotEmpty } from 'class-validator';
-
-export class DeleteBoardDto {
-  @IsNotEmpty()
-  userId: number;
-
-  @IsNotEmpty()
-  boardId: number;
-}
