+++ server/src/board/dto/like-board.dto.ts
@@ -1,19 +1,14 @@
 import { IsNotEmpty, IsInt } from 'class-validator';
 import { Transform } from 'class-transformer';
 
-export class GetCommentsDto {
+export class LikeBoardDto {
   @IsNotEmpty()
   @IsInt()
   @Transform(({ value }) => parseInt(value, 10))
-  board_id: number;
+  userId: number;
 
   @IsNotEmpty()
   @IsInt()
   @Transform(({ value }) => parseInt(value, 10))
-  count: number;
-
-  @IsNotEmpty()
-  @IsInt()
-  @Transform(({ value }) => parseInt(value, 10))
-  max_id: number;
+  boardId: number;
 }
