--- server/src/board/dto/likeBoardDto.ts
@@ -1,9 +1,0 @@
-import { IsNotEmpty } from 'class-validator';
-
-export class LikeBoardDto {
-  @IsNotEmpty()
-  userId: number;
-
-  @IsNotEmpty()
-  boardId: number;
-}
