--- server/src/board/dto/updateBoaedDto.ts
@@ -1,12 +1,0 @@
-import { IsNotEmpty } from 'class-validator';
-
-export class UpdateBoardDto {
-  @IsNotEmpty()
-  userId: number;
-
-  @IsNotEmpty()
-  boardId: number;
-
-  @IsNotEmpty()
-  content: string;
-}
