+++ server/src/comment/dto/delete-comment.dto.ts
@@ -1,9 +1,14 @@
-import { IsNotEmpty } from 'class-validator';
+import { IsNotEmpty, IsInt } from 'class-validator';
+import { Transform } from 'class-transformer';
 
 export class DeleteCommentDto {
   @IsNotEmpty()
+  @IsInt()
+  @Transform(({ value }) => parseInt(value, 10))
   boardId: number;
 
   @IsNotEmpty()
+  @IsInt()
+  @Transform(({ value }) => parseInt(value, 10))
   userId: number;
 }
