+++ server/src/comment/dto/update-comment.dto.ts
@@ -1,10 +1,10 @@
-import { IsNotEmpty } from 'class-validator';
+import { Transform } from 'class-transformer';
+import { IsInt, IsNotEmpty } from 'class-validator';
 
 export class UpdateCommentDto {
   @IsNotEmpty()
-  boardId: number;
-
-  @IsNotEmpty()
+  @IsInt()
+  @Transform(({ value }) => parseInt(value, 10))
   userId: number;
 
   @IsNotEmpty()
